pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-anomaly-detection"
version = "0.1.0"
description = "Credit card fraud detection project with rule-based, ML, Kafka, and Flink components."
requires-python = ">=3.10"
dependencies = [
    "imbalanced-learn>=0.12",
    "joblib>=1.3",
    "kaggle>=2.0.0",
    "kafka-python>=2.0",
    "matplotlib>=3.8",
    "numpy>=1.26",
    "pandas>=2.1",
    "scikit-learn>=1.4",
    "seaborn>=0.13",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
